% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inteRplot.R
\name{inteRplot}
\alias{inteRplot}
\title{inteRplot}
\usage{
inteRplot(
  model,
  pred,
  mod,
  outcome = "outcome",
  xaxis = "predictor",
  moderator = "moderator",
  miny = 1,
  maxy = 5,
  xlabels = c("Low", "High")
)
}
\arguments{
\item{model}{a vector of your variables}

\item{pred}{a vector of your variables}

\item{mod}{a vector of your variables}

\item{outcome}{a vector of your variables}

\item{xaxis}{a vector of your variables}

\item{moderator}{a vector of your variables}

\item{miny}{a vector of your variables}

\item{maxy}{a vector of your variables}

\item{xlabels}{a vector of your variables}
}
\value{
interaction figure
}
\description{
informative interaction figure
}
