% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_diagnostics.R
\name{lm.diagnostics}
\alias{lm.diagnostics}
\title{lm.diagnostics}
\usage{
lm.diagnostics(model)
}
\arguments{
\item{model}{your lm model}
}
\value{
Six plot with each a different method to check the model diagnostics
}
\description{
This function plots an overview of six different methods to check model diagnostics.
}
