% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manovaR.R
\name{manovaR}
\alias{manovaR}
\title{manovaR}
\usage{
manovaR(variables, group, dataset)
}
\arguments{
\item{variables}{a vector of your variables}

\item{group}{a character of the name of your group variable}

\item{dataset}{name of the dataset}
}
\value{
MANOVA table with descriptives and statistics
}
\description{
manova table with statistics
}
