% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicoR.R
\name{multicoR}
\alias{multicoR}
\title{multicoR}
\usage{
multicoR(dataset)
}
\arguments{
\item{dataset}{Your dataset with the variable in order you want to present them}
}
\value{
A html correlation matrix with significance stars, Note and descriptive statistics
}
\description{
This function plots a correlation table in APA style with descriptive statistics.
}
