% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteRs.R
\name{clusteRs}
\alias{clusteRs}
\title{clusteRs}
\usage{
clusteRs(dataset)
}
\arguments{
\item{dataset}{dataset}
}
\value{
overview of indices for number of clusters
}
\description{
overview of indices for number of clusters
}
