% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manovaR.R
\name{manovaR}
\alias{manovaR}
\title{manovaR}
\usage{
manovaR(variables, group, dataset)
}
\arguments{
\item{model}{your lm model}
}
\value{
Six plot with each a different method to check the model diagnostics
}
\description{
manova table with statistics
}
