% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manovaR.R
\name{manovaR}
\alias{manovaR}
\title{manovaR}
\usage{
manovaR(data, tukey = FALSE, sign = 0.05)
}
\arguments{
\item{data}{Your data with the variable in order you want to present them}

\item{tukey}{Your data with the variable in order you want to present them}

\item{sign}{significance level of multi comparison}
}
\value{
A html correlation matrix with significance stars, Note and descriptive statistics
}
\description{
This function plots a correlation table in APA style with descriptive statistics.
}
